Added functionality to return the results of queries.
I have updated the executeQuery() method to now return the results of queries when specified using a boolean flag.
This allows data to be fetched from the tables in the database using SELECT (or other) queries and then manipulated.

I have also changed the name of the loginInfo table from 'loginInfo' to 'User' as this better represents the data the table stores.