Created a file for populating the database with test data.
I have created a .json file which stores the data about the menu items that will be displayed on the website.
So far the file contains only one object, and by running the command "python -m manage.py loaddata menuData.json",
the data will be copied from this file into the corresponding tables of the database, so it can be used for testing.