Fix notify order issue when not logged in
There was an error that would pop up if a a logged out user tries to
access the home page, this is because there orders were trying to be
accessed even though they dont have any.

To fix this i check if a user is logged in first before trying to access
there orders.